+++ projects/mat-video/src/lib/pipes/seconds-to-time.pipe.ts
@@ -1,7 +1,7 @@
-import { Pipe, PipeTransform } from "@angular/core";
+import { Pipe, PipeTransform } from '@angular/core';
 
 @Pipe({
-  name: "secondsToTime"
+  name: 'secondsToTime'
 })
 export class SecondsToTimePipe implements PipeTransform {
   times = {
@@ -13,19 +13,19 @@ export class SecondsToTimePipe implements PipeTransform {
 
   transform(seconds: number) {
     if (!seconds) {
-      return "0:00";
+      return '0:00';
     } else {
-      let timeString = "";
+      let timeString = '';
       for (const key in this.times) {
         if (Math.floor(seconds / this.times[key]) > 0) {
-          timeString += Math.floor(seconds / this.times[key]).toString() + ":";
+          timeString += Math.floor(seconds / this.times[key]).toString() + ':';
           seconds = seconds - this.times[key] * Math.floor(seconds / this.times[key]);
         }
       }
-      timeString += Math.floor(seconds / 60).toString() + ":";
+      timeString += Math.floor(seconds / 60).toString() + ':';
       seconds = seconds - 60 * Math.floor(seconds / 60);
       if (Math.floor(seconds) < 10) {
-        timeString += "0";
+        timeString += '0';
       }
       timeString += Math.floor(seconds).toString();
       return timeString;