+++ projects/mat-video/src/lib/ui/mat-download-button/mat-download-button.component.ts
@@ -1,9 +1,9 @@
-import { Component, OnInit, Input } from "@angular/core";
+import { Component, Input } from '@angular/core';
 
 @Component({
-  selector: "mat-download-button",
-  templateUrl: "./mat-download-button.component.html",
-  styleUrls: ["./mat-download-button.component.scss"]
+  selector: 'mat-download-button',
+  templateUrl: './mat-download-button.component.html',
+  styleUrls: ['./mat-download-button.component.scss']
 })
 export class MatDownloadButtonComponent {
   @Input() video: HTMLVideoElement;