+++ projects/mat-video/src/lib/ui/mat-frame-by-frame-control/mat-frame-by-frame-control.component.ts
@@ -1,9 +1,9 @@
-import { Component, Input, OnInit } from "@angular/core";
+import { Component, Input, OnInit } from '@angular/core';
 
 @Component({
-  selector: "mat-frame-by-frame-control",
-  templateUrl: "./mat-frame-by-frame-control.component.html",
-  styleUrls: ["./mat-frame-by-frame-control.component.scss"]
+  selector: 'mat-frame-by-frame-control',
+  templateUrl: './mat-frame-by-frame-control.component.html',
+  styleUrls: ['./mat-frame-by-frame-control.component.scss']
 })
 export class MatFrameByFrameControlComponent implements OnInit {
   @Input() video: HTMLVideoElement;