+++ projects/mat-video/src/lib/ui/mat-play-button/mat-play-button.component.ts
@@ -1,12 +1,12 @@
-import { AfterViewInit, Component, EventEmitter, HostListener, Input, OnDestroy, Output, Renderer2 } from "@angular/core";
+import { AfterViewInit, Component, EventEmitter, HostListener, Input, OnDestroy, Output, Renderer2 } from '@angular/core';
 
-import { EventHandler } from "../../interfaces/event-handler.interface";
-import { EventService } from "../../services/event.service";
+import { EventHandler } from '../../interfaces/event-handler.interface';
+import { EventService } from '../../services/event.service';
 
 @Component({
-  selector: "mat-play-button",
-  templateUrl: "./mat-play-button.component.html",
-  styleUrls: ["./mat-play-button.component.scss"]
+  selector: 'mat-play-button',
+  templateUrl: './mat-play-button.component.html',
+  styleUrls: ['./mat-play-button.component.scss']
 })
 export class MatPlayButtonComponent implements AfterViewInit, OnDestroy {
   @Input() video: HTMLVideoElement;
@@ -23,11 +23,11 @@ export class MatPlayButtonComponent implements AfterViewInit, OnDestroy {
 
   ngAfterViewInit(): void {
     this.events = [
-      { element: this.video, name: "play", callback: event => this.setVideoPlayback(true), dispose: null },
-      { element: this.video, name: "pause", callback: event => this.setVideoPlayback(false), dispose: null },
-      { element: this.video, name: "durationchange", callback: event => this.setVideoPlayback(false), dispose: null },
-      { element: this.video, name: "ended", callback: event => this.setVideoPlayback(false), dispose: null },
-      { element: this.video, name: "click", callback: event => this.toggleVideoPlayback(), dispose: null }
+      { element: this.video, name: 'play', callback: event => this.setVideoPlayback(true), dispose: null },
+      { element: this.video, name: 'pause', callback: event => this.setVideoPlayback(false), dispose: null },
+      { element: this.video, name: 'durationchange', callback: event => this.setVideoPlayback(false), dispose: null },
+      { element: this.video, name: 'ended', callback: event => this.setVideoPlayback(false), dispose: null },
+      { element: this.video, name: 'click', callback: event => this.toggleVideoPlayback(), dispose: null }
     ];
 
     this.evt.addEvents(this.renderer, this.events);
@@ -53,7 +53,7 @@ export class MatPlayButtonComponent implements AfterViewInit, OnDestroy {
     this.playChanged.emit(this.play);
   }
 
-  @HostListener("document:keyup.space", ["$event"])
+  @HostListener('document:keyup.space', ['$event'])
   onPlayKey(event: KeyboardEvent) {
     if (this.keyboard) {
       this.toggleVideoPlayback();