+++ projects/mat-video/src/lib/ui/mat-quality-control/mat-quality-control.component.ts
@@ -1,9 +1,9 @@
-import { Component, Input } from "@angular/core";
+import { Component, Input } from '@angular/core';
 
 @Component({
-  selector: "mat-quality-control",
-  templateUrl: "./mat-quality-control.component.html",
-  styleUrls: ["./mat-quality-control.component.scss"]
+  selector: 'mat-quality-control',
+  templateUrl: './mat-quality-control.component.html',
+  styleUrls: ['./mat-quality-control.component.scss']
 })
 export class MatQualityControlComponent {
   @Input() video: HTMLVideoElement;