+++ projects/mat-video/src/lib/ui/mat-time-control/mat-time-control.component.ts
@@ -1,9 +1,9 @@
-import { Component, Input } from "@angular/core";
+import { Component, Input } from '@angular/core';
 
 @Component({
-  selector: "mat-time-control",
-  templateUrl: "./mat-time-control.component.html",
-  styleUrls: ["./mat-time-control.component.scss"]
+  selector: 'mat-time-control',
+  templateUrl: './mat-time-control.component.html',
+  styleUrls: ['./mat-time-control.component.scss']
 })
 export class MatTimeControlComponent {
   @Input() video: HTMLVideoElement;