+++ projects/mat-video/src/lib/ui/mat-video-spinner/mat-video-spinner.component.ts
@@ -1,22 +1,22 @@
-import { AfterViewInit, Component, Input, OnDestroy, Renderer2 } from "@angular/core";
+import { AfterViewInit, Component, Input, OnDestroy, Renderer2 } from '@angular/core';
 
-import { EventHandler } from "../../interfaces/event-handler.interface";
-import { EventService } from "../../services/event.service";
+import { EventHandler } from '../../interfaces/event-handler.interface';
+import { EventService } from '../../services/event.service';
 
 @Component({
-  selector: "mat-video-spinner",
-  templateUrl: "./mat-video-spinner.component.html",
+  selector: 'mat-video-spinner',
+  templateUrl: './mat-video-spinner.component.html',
   styleUrls: [
-    "./mat-video-spinner.component.scss",
-    "./indicators/spin.scss",
-    "./indicators/dot.scss",
-    "./indicators/split-ring.scss",
-    "./indicators/hourglass.scss"
+    './mat-video-spinner.component.scss',
+    './indicators/spin.scss',
+    './indicators/dot.scss',
+    './indicators/split-ring.scss',
+    './indicators/hourglass.scss'
   ]
 })
 export class MatVideoSpinnerComponent implements AfterViewInit, OnDestroy {
   @Input() video: HTMLVideoElement;
-  @Input() spinner = "spin";
+  @Input() spinner = 'spin';
 
   videoBuffering = false;
   videoLoaded = false;
@@ -27,11 +27,11 @@ export class MatVideoSpinnerComponent implements AfterViewInit, OnDestroy {
 
   ngAfterViewInit(): void {
     this.events = [
-      { element: this.video, name: "loadstart", callback: event => (this.videoLoaded = false), dispose: null },
-      { element: this.video, name: "loadedmetadata", callback: event => (this.videoLoaded = true), dispose: null },
-      { element: this.video, name: "canplay", callback: event => (this.videoBuffering = false), dispose: null },
-      { element: this.video, name: "waiting", callback: event => (this.videoBuffering = true), dispose: null },
-      { element: this.video, name: "durationchange", callback: event => (this.videoBuffering = true), dispose: null }
+      { element: this.video, name: 'loadstart', callback: event => (this.videoLoaded = false), dispose: null },
+      { element: this.video, name: 'loadedmetadata', callback: event => (this.videoLoaded = true), dispose: null },
+      { element: this.video, name: 'canplay', callback: event => (this.videoBuffering = false), dispose: null },
+      { element: this.video, name: 'waiting', callback: event => (this.videoBuffering = true), dispose: null },
+      { element: this.video, name: 'durationchange', callback: event => (this.videoBuffering = true), dispose: null }
     ];
 
     this.video.onloadeddata = () => (this.videoLoaded = true);