+++ projects/mat-video/src/lib/ui/mat-volume-control/mat-volume-control.component.ts
@@ -1,37 +1,39 @@
-import { Component, EventEmitter, HostListener, Input, Output } from "@angular/core";
-import { ThemePalette } from "@angular/material/core";
+import { AfterViewInit, Component, EventEmitter, HostListener, Input, OnChanges, Output, SimpleChanges } from '@angular/core';
+import { ThemePalette } from '@angular/material/core';
 
-import { EventService } from "../../services/event.service";
+import { EventService } from '../../services/event.service';
 
 @Component({
-  selector: "mat-volume-control",
-  templateUrl: "./mat-volume-control.component.html",
-  styleUrls: ["./mat-volume-control.component.scss"]
+  selector: 'mat-volume-control',
+  templateUrl: './mat-volume-control.component.html',
+  styleUrls: ['./mat-volume-control.component.scss']
 })
-export class MatVolumeControlComponent {
+export class MatVolumeControlComponent implements AfterViewInit, OnChanges {
   @Input() video: HTMLVideoElement = null;
 
-  @Input() color: ThemePalette = "primary";
+  @Input() color: ThemePalette = 'primary';
 
   @Input() volume = 1;
 
   @Output() volumeChanged = new EventEmitter<number>();
 
-  private pMuted = false;
-  @Input()
-  get muted() {
-    return this.pMuted;
-  }
-  set muted(v: boolean) {
-    this.pMuted = v;
-    this.video.muted = this.pMuted;
-  }
+  @Input() muted = false;
 
   @Output() mutedChanged = new EventEmitter<boolean>();
 
   @Input() keyboard = true;
 
   constructor(private evt: EventService) {}
+
+  ngAfterViewInit(): void {
+    this.updateMuted(false);
+  }
+
+  ngOnChanges(changes: SimpleChanges): void {
+    if (changes.muted) {
+      this.updateMuted(false);
+    }
+  }
 
   setVolume(value: number): void {
     this.volume = value;
@@ -54,12 +56,17 @@ export class MatVolumeControlComponent {
     this.updateMuted();
   }
 
-  updateMuted(): void {
-    this.video.muted = this.muted;
-    this.mutedChanged.emit(this.muted);
+  updateMuted(emitChange: boolean = true): void {
+    if (this.video) {
+      this.video.muted = this.muted;
+    }
+
+    if (emitChange) {
+      this.mutedChanged.emit(this.muted);
+    }
   }
 
-  @HostListener("document:keyup.m", ["$event"])
+  @HostListener('document:keyup.m', ['$event'])
   onMuteKey(event: KeyboardEvent) {
     if (this.keyboard) {
       this.toggleMuted();