+++ projects/mat-video/src/public-api.ts
@@ -2,4 +2,4 @@
  * Public API Surface of mat-video
  */
 
-export * from "./lib/video.module";
+export * from './lib/video.module';